+++ experimental/ssh/internal/setup/setup.go
@@ -2,6 +2,7 @@ package setup
 
 import (
 	"context"
+	"errors"
 	"fmt"
 	"os"
 	"path/filepath"
@@ -20,6 +21,8 @@ type SetupOptions struct {
 	HostName string
 	// The cluster ID to connect to
 	ClusterID string
+	// Whether to automatically start the cluster during ssh connection if it is not running
+	AutoStartCluster bool
 	// Delay before shutting down the SSH tunnel, will be added as a --shutdown-delay flag to the ProxyCommand
 	ShutdownDelay time.Duration
 	// Optional path to the local ssh config. Defaults to ~/.ssh/config
@@ -74,8 +77,8 @@ Host %s
     ConnectTimeout 360
     StrictHostKeyChecking accept-new
     IdentityFile %q
-    ProxyCommand %q ssh connect --proxy --cluster=%s --shutdown-delay=%s %s
-`, opts.HostName, identityFilePath, execPath, opts.ClusterID, opts.ShutdownDelay, profileOption)
+    ProxyCommand %q ssh connect --proxy --cluster=%s --auto-start-cluster=%t --shutdown-delay=%s %s
+`, opts.HostName, identityFilePath, execPath, opts.ClusterID, opts.AutoStartCluster, opts.ShutdownDelay, profileOption)
 
 	return hostConfig, nil
 }
@@ -141,7 +144,38 @@ func updateSSHConfigFile(configPath, hostConfig, hostName string) error {
 	return nil
 }
 
+func clusterSelectionPrompt(ctx context.Context, client *databricks.WorkspaceClient) (string, error) {
+	spinnerChan := cmdio.Spinner(ctx)
+	spinnerChan <- "Loading clusters."
+	clusters, err := client.Clusters.ClusterDetailsClusterNameToClusterIdMap(ctx, compute.ListClustersRequest{
+		FilterBy: &compute.ListClustersFilterBy{
+			ClusterSources: []compute.ClusterSource{compute.ClusterSourceApi, compute.ClusterSourceUi},
+		},
+	})
+	close(spinnerChan)
+	if err != nil {
+		return "", fmt.Errorf("failed to load names for Clusters drop-down. Please manually specify cluster argument. Original error: %w", err)
+	}
+	id, err := cmdio.Select(ctx, clusters, "The cluster to connect to")
+	if err != nil {
+		return "", err
+	}
+	return id, nil
+}
+
 func Setup(ctx context.Context, client *databricks.WorkspaceClient, opts SetupOptions) error {
+	if opts.ClusterID == "" {
+		id, err := clusterSelectionPrompt(ctx, client)
+		if err != nil {
+			return err
+		}
+		opts.ClusterID = id
+	}
+
+	if opts.ClusterID == "" {
+		return errors.New("cluster ID is required")
+	}
+
 	err := validateClusterAccess(ctx, client, opts.ClusterID)
 	if err != nil {
 		return err