+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Version changelog
 
+## Release v0.269.0
+
+### Bundles
+* Fix checkForPreventDestroy to check all resources if some does not have prevent_destroy set ([#3615](https://github.com/databricks/cli/pull/3615))
+
+
 ## Release v0.268.0
 
 ### Bundles
@@ -24,6 +30,7 @@
 * Add support for specifying database instance as an application resource ([#3529](https://github.com/databricks/cli/pull/3529))
 * Allow referencing job libraries outside bundle root without the need to specify sync root ([#2842](https://github.com/databricks/cli/pull/2842))
 * Add top level `run_as` support for Lakeflow Declarative Pipelines ([#3307](https://github.com/databricks/cli/pull/3307))
+* Add support for `lifecycle prevent_destroy` option ([#3448](https://github.com/databricks/cli/pull/3448))
 
 
 ## Release v0.266.0